Add preliminary work for a yaml based configuration file.
I am temporarily shelving this because it is a lot of work, and I could
probably get away with something much, much more simple that would work
just as well, and is also built in to Flask, using the Configure object.

I'll come back to this later if I feel it would be better.